feat: 위치 인사이트 버블 차트 추가 및 카테고리 한글 매핑
- BubbleChartView: CALayer 기반 버블 차트 컴포넌트 구현
- InsightLocationStatsView: 위치 통계 카드 뷰 추가
- InsightCategoryStatsView: FoodGenre.displayName으로 한글 매핑
- PhotoStatsDTO.changeRate: Double 타입으로 변경하여 디코딩 오류 수정

+++ FoodDiary/Data/Sources/DTO/InsightResponseDTO.swift
@@ -44,7 +44,7 @@ struct InsightResponseDTO: Decodable {
 struct PhotoStatsDTO: Decodable {
     let currentMonthCount: Int
     let previousMonthCount: Int
-    let changeRate: Int
+    let changeRate: Double
 
     enum CodingKeys: String, CodingKey {
         case currentMonthCount = "current_month_count"
@@ -56,7 +56,7 @@ struct PhotoStatsDTO: Decodable {
         PhotoStats(
             currentMonthCount: currentMonthCount,
             previousMonthCount: previousMonthCount,
-            changeRate: changeRate
+            changeRate: Int(changeRate)
         )
     }
 }
